seller/requirements.txt
Flask-PyMongo==2.3.0
PyMySQL==1.1.0
